Now handles SPAligner
