+++ solve_with_long_reads.py
@@ -10,7 +10,8 @@ import re #to find all numbers in a mixed number/letters string (such as 31M1D4M
 from collections import Counter #to count the elements in a list quickly
 from scipy import sparse #to handle sparse matrices
 from determine_multiplicity import determine_multiplicity
-
+from input_output import read_GAF
+from input_output import read_TSV
 
 import segment
 
@@ -51,9 +52,17 @@ def bridge_with_long_reads(segments, names, copiesnumber, gafFile, supported_lin
     supported_links = sparse.lil_matrix((len(names)*2, len(names)*2)) #supported links is the list of the links between different contigs found in the gaf file
     
     lines = []
-    print("Reading the gaf file...")
-    read_GAF(gafFile, 0.7, 0.1, lines)
-    print("Finished going through the gaf file.")
+    if '.gaf' in gafFile :
+        print("Reading the gaf file...")
+        read_GAF(gafFile, 0.7, 0.1, lines)
+        print("Finished going through the gaf file.")
+    elif '.tsv' in gafFile :
+        print("Reading the gpa file...")
+        read_TSV(gafFile, names, lines)
+        print("Finished going through the gpa file.")
+    else :
+        print("ERROR: input format of mapped read not recognized. It should be .gfa or .gpa")
+        sys.exit()
     
     sure_haploids = False
     
@@ -111,25 +120,7 @@ def bridge_with_long_reads(segments, names, copiesnumber, gafFile, supported_lin
     unzip_graph_with_bridges(segments, non_overlapping_bridges, copiesnumber, haploidContigs, haploidContigsNames, names, supported_links2, minimum_supported_links, multiplicities, longContigs)
         
     #print(non_overlapping_bridges)
-
-#input : GAF file and parameters telling which line are deemed informative
-#output : list of useful lines extracted (['>12>34<2' , '>77<33' ,... ] for example)
-def read_GAF(gafFile,similarity_threshold, whole_mapping_threshold, lines) : #a function going through the gaf files and inventoring all useful lines
-    
-    gaf = open(gafFile, 'r')
-    
-    for line in gaf :
-        ls = line.split('\t')
-        path = ls[5] # in GAF format, the 6th column is the path on which the read matched
-        
-
-        if ls[5].count('>') + ls[5].count('<') > 1 :
-                        
-            if (not 'id:f' in ls[-2]) or (float(ls[-2].split(':')[-1]) > similarity_threshold) :
-                
-                if (float(ls[3])-float(ls[2]))/float(ls[1]) > whole_mapping_threshold :
-    
-                    lines += [ls[5]]              
+            
                   
 #input : a list of alignments of a gaf file
 #output : the completed bridges list, with for each haploid contig a list of what was found left and right of the contig
@@ -315,7 +306,6 @@ def compute_supported_links(supported_links, consensus_bridges, haploidContigsNa
 #output: a list of non-overlapping bridges left and right of each contig, with only full bridges (connecting 2 haploid contigs)
 def merge_bridges(non_overlapping_bridges, consensus_bridges, haploidContigsNames, haploidContigs, longContigs, names) :
     
-    print("Merging bridges")
     sure_haploids = True
     not_actually_haploid = [] #a list of not actually haploid contigs among the haploidContigs
     
@@ -383,9 +373,7 @@ def merge_bridges(non_overlapping_bridges, consensus_bridges, haploidContigsName
             reliable_haploid_contigs += [haploidContigs[i]]
             reliable_haploid_contigsNames[haploidContigs[i].names[0]] = index
             index += 1
-            
-    print("Done merging bridges")
-        
+                    
     return reliable_haploid_contigs, reliable_haploid_contigsNames, new_consensus_bridges, sure_haploids #return the updated list of haploid contigs
